// src/app/enums.ts
export enum Configurations {
  ACCESS_TOKEN = 'ACCESS_TOKEN',
  SEARCH_CRITERIA = 'SEARCH_CRITERIA',
  USER_DATA = 'USER DATA'
}

export enum ConfigurationTypes {
  Colors = 'Color',
  CarTypes = 'CarType',
  Features = 'Feature',
  ExtraOptions = 'ExtraOption',
  FuelTypes = 'FuelType',
  GpsTypes = 'GpsType',
  Countries = 'Country',
  ImageTags = 'ImageTag',
  RentalPeriods = 'RentalPeriodType',
  Languages = 'Language',
  Years = 'Year',
  InsuranceTypes = 'InsuranceType',
  TransmissionTypes = 'TransmissionType',
  VehicleCategories = 'VehicleCategory',
  AssetMaintenance = 'AssetMaintenance',
  RoadsideAssistance = 'RoadsideAssistance',
  PricingTypes = 'PricingType',
  Currencies = 'Currency',
  TimeZone = 'TimeZone',
  DecimalPlace = 'DecimalPlace',
  LongDateFormat = 'LongDateFormat',
  ShortDateFormat = 'ShortDateFormat',
  Unit = 'Unit',
  CustomerTypes = 'CustomerType',
  RentalCategories = 'RentalCategory',
  RentalTypes = 'RentalType',
  RentChannels = 'RentChannel',
  RentChannelRefs = 'RentChannelRef',
  RenterTypes = 'RenterType',
  OfficeTypes = 'OfficeType',
  PaymentModes = 'PaymentMode',
  AdvanceNotice = 'AdvanceNotice',
  AssetUnavailabilityReason = 'AssetUnavailabilityReason'
}

export enum RentalCategories {
  CAR = 'Car',
  CARWITHDRIVER = 'CarWithDriver',
  SUBSCRIPTION = 'Subscription',
  AIRPORTTRANSFER = 'AirportTransfer'
}

export enum LookupType {
  ASSET_MANAGEMENT = 'Asset_Management',
  PLATFORM_MANAGEMENT = 'Platform_Management',
  BOOKING_MANAGEMENT = 'Booking_Management',
  LANGUAGE = 'Language'
}

export enum LanguageType {
  ENGLISH = 'en',
  THAI = 'th',
  CHINESE = 'zh',
  CHINESE_FULL = 'zh-cn'
}